Fix ping reactivating when changing target
This fixes the ping being reactivated when initiating a reply to a message while the reply bar is already showing. This happens when you change the message you wanna reply to without explicitly canceling the reply. In fact even right-click + reply on the same message you are already replying to would cause the ping to reactivate.

I've also updated the source url (and updateUrl, not that it would matter) to point to this repo instead, as the upstream repo is behind and isn't updating anymore. Didn't really feel comfortable with changing the author too.